Alternative
I am attempting to follow the instructions on this page to use this SDK to test the Pubsub emulator, but the requirements.txt file it references looks to have an unresolved conflict in here. The commit touching this file claims to have resolved a host of conflicts, but it looks like this file got missed.